flaskserver/requirements.txt
Flask
psycopg2-binary
python-dotenv
paho-mqtt